Homework2
