--- app/src/main/java/com/szareckii/nasapictureoftheday/ui/picture/fragment/ChipsFragment.kt
@@ -1,32 +1,0 @@
-package com.szareckii.nasapictureoftheday.ui.picture.fragment
-
-import android.os.Bundle
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
-import android.widget.Toast
-import androidx.fragment.app.Fragment
-import com.google.android.material.chip.Chip
-import com.szareckii.nasapictureoftheday.R
-import kotlinx.android.synthetic.main.fragment_chips.*
-
-class ChipsFragment: Fragment() {
-
-    override fun onCreateView(
-        inflater: LayoutInflater, container: ViewGroup?,
-        savedInstanceState: Bundle?
-    ): View? {
-        return inflater.inflate(R.layout.fragment_chips, container, false)
-    }
-
-    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
-        super.onViewCreated(view, savedInstanceState)
-        chipGroup.setOnCheckedChangeListener { chipGroup, position ->
-            chipGroup.findViewById<Chip>(position)?.let {
-                Toast.makeText(context, "Выбран ${it.text}", Toast.LENGTH_SHORT).show()
-            }
-        }
-    }
-
-
-}
