+++ app/src/main/java/com/szareckii/nasapictureoftheday/ui/picture/fragment/PictureOfTheDayFragment.kt
@@ -67,10 +67,9 @@ class PictureOfTheDayFragment : Fragment() {
 
     override fun onOptionsItemSelected(item: MenuItem): Boolean {
         when (item.itemId) {
-            R.id.app_bar_fav -> toast("Favourite")
             R.id.app_bar_search -> toast("Search")
             R.id.app_bar_settings -> activity?.supportFragmentManager?.beginTransaction()?.add(
-                R.id.container, ChipsFragment()
+                R.id.container, SettingsFragment()
             )?.addToBackStack(null)?.commit()
 
             android.R.id.home -> {
