+++ app/src/main/java/com/szareckii/nasapictureoftheday/ui/picture/fragment/SettingsFragment.kt
@@ -1,0 +1,51 @@
+package com.szareckii.nasapictureoftheday.ui.picture.fragment
+
+import android.content.Context
+import android.content.SharedPreferences
+import android.os.Bundle
+import android.view.Gravity
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import android.widget.Toast
+import androidx.fragment.app.Fragment
+import com.google.android.material.chip.Chip
+import com.szareckii.nasapictureoftheday.R
+import kotlinx.android.synthetic.main.fragment_settings.*
+
+class SettingsFragment: Fragment() {
+
+    override fun onCreateView(
+        inflater: LayoutInflater, container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View? {
+        return inflater.inflate(R.layout.fragment_settings, container, false)
+    }
+
+    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
+        super.onViewCreated(view, savedInstanceState)
+
+        val sharedPref = activity?.getPreferences(Context.MODE_PRIVATE) ?: return
+
+        chipGroup.setOnCheckedChangeListener { chipGroup, position ->
+            chipGroup.findViewById<Chip>(position)?.let {
+                saveSharedPref(sharedPref, it.text.toString())
+            }
+        }
+    }
+
+    private fun saveSharedPref(sharedPref: SharedPreferences, text: String) {
+        if (text == getString(R.string.theme_day)) {
+                with (sharedPref.edit()) {
+                    putInt(getString(R.string.theme), 1)
+                    apply()
+                }
+        } else {
+                with (sharedPref.edit()) {
+                    putInt(getString(R.string.theme), 2)
+                    apply()
+            }
+        }
+        activity?.recreate()
+    }
+}
