+++ app/src/main/res/color/text_field_background.xml
@@ -1,0 +1,6 @@
+<?xml version="1.0" encoding="utf-8"?>
+<selector xmlns:android="http://schemas.android.com/apk/res/android">
+    <item android:alpha="0.16" android:color="?attr/colorPrimary" android:state_hovered="true"/>
+    <item android:alpha="0.04" android:color="?attr/colorPrimary" android:state_enabled="false"/>
+    <item android:alpha="0.12" android:color="?attr/colorPrimary"/>
+</selector>
