+++ app/src/main/res/layout/bottom_navigation_layout.xml
@@ -5,12 +5,10 @@
 
     <com.google.android.material.navigation.NavigationView
         android:id="@+id/navigation_view"
+        style="@style/BottomNavigation"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
         android:layout_gravity="bottom"
-        android:layout_marginStart="8dp"
-        android:layout_marginTop="8dp"
-        android:layout_marginEnd="8dp"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toTopOf="parent"
