+++ app/src/main/res/menu/menu_bottom_bar.xml
@@ -2,14 +2,8 @@
     xmlns:app="http://schemas.android.com/apk/res-auto">
 
     <item
-        android:id="@+id/app_bar_fav"
-        android:icon="@drawable/ic_favourite_menu"
-        android:title="@string/favourite"
-        app:showAsAction="ifRoom" />
-
-    <item
-        android:id="@+id/app_bar_search"
-        android:icon="@drawable/ic_search"
+        android:id="@+id/app_bar_settings"
+        android:icon="@drawable/ic_settings"
         android:title="@string/settings"
         app:showAsAction="ifRoom" />
 </menu>
