+++ app/src/main/res/menu/menu_bottom_bar_other_screen.xml
@@ -3,8 +3,8 @@
     xmlns:app="http://schemas.android.com/apk/res-auto">
 
     <item
-        android:id="@+id/app_bar_settings"
-        android:icon="@drawable/ic_settings_menu"
-        android:title="@string/search"
+        android:id="@+id/app_bar_search"
+        android:icon="@drawable/ic_search"
+        android:title="@string/help"
         app:showAsAction="ifRoom" />
 </menu>
