+++ app/src/main/res/values/attr.xml
@@ -1,0 +1,4 @@
+<?xml version="1.0" encoding="utf-8"?>
+<resources>
+    <attr name="textColorEditText" format="reference"/>
+</resources>
