+++ app/src/main/res/values/colors.xml
@@ -13,4 +13,9 @@
     <color name="teal_700">#FF018786</color>
     <color name="black">#FF000000</color>
     <color name="white">#FFFFFFFF</color>
-</resources>
+
+    <color name="wisteria">#66D8A7E5</color>
+    <color name="whisper">#66f6f6fa</color>
+    <color name="bright_gray">#663b4351</color>
+
+    </resources>
