+++ app/src/main/res/values/dimension.xml
@@ -1,0 +1,10 @@
+<?xml version="1.0" encoding="utf-8"?>
+<resources>
+    <dimen name="margin_small">8dp</dimen>
+    <dimen name="behavior_peekHeight">150dp</dimen>
+    <dimen name="margin_top_small">5dp</dimen>
+    <dimen name="text_size_big">20sp</dimen>
+    <dimen name="padding">10dp</dimen>
+    <dimen name="paddingStart">10dp</dimen>
+    <dimen name="paddingEnd">10dp</dimen>
+</resources>
